# Optimize binary `get_number` implementation by reading multiple bytes at once