@nlohmann 2024-11-26
I like this idea! @TianyiChen any plans to continue working on this?