@TianyiChen 2024-11-27
Sure, I have rebased the branch and added a from msgpack test using C FILE. Could you retrigger the CI and let's see what's missing right now?