@nlohmann 2024-11-28
I could reproduce the benchmark results. Good job, thank you!

### Before

```
FromMsgpack/jeopardy          1149154833 ns   1148401000 ns            1 bytes_per_second=38.3318Mi/s
FromMsgpack/canada              26770413 ns     26757962 ns           26 bytes_per_second=37.6418Mi/s
FromMsgpack/citm_catalog         9823588 ns      9819069 ns           72 bytes_per_second=33.2626Mi/s
FromMsgpack/twitter             10075707 ns     10066348 ns           69 bytes_per_second=38.0386Mi/s
FromMsgpack/floats             175452521 ns    175401500 ns            4 bytes_per_second=48.9339Mi/s
FromMsgpack/signed_ints        176531916 ns    176491750 ns            4 bytes_per_second=48.6316Mi/s
FromMsgpack/unsigned_ints      176015917 ns    175288250 ns            4 bytes_per_second=48.9655Mi/s
FromMsgpack/small_signed_ints  108130396 ns    107524500 ns            6 bytes_per_second=43.4686Mi/s
```

### After

```
FromMsgpack/jeopardy          1137093749 ns   1134339000 ns            1 bytes_per_second=38.807Mi/s
FromMsgpack/canada              13543979 ns     13528462 ns           52 bytes_per_second=74.4518Mi/s
FromMsgpack/citm_catalog         8898435 ns      8895269 ns           78 bytes_per_second=36.717Mi/s
FromMsgpack/twitter              9764953 ns      9760380 ns           71 bytes_per_second=39.231Mi/s
FromMsgpack/floats              58513708 ns     58492750 ns           12 bytes_per_second=146.737Mi/s
FromMsgpack/signed_ints         59872042 ns     59853500 ns           12 bytes_per_second=143.401Mi/s
FromMsgpack/unsigned_ints       55908673 ns     55896000 ns           13 bytes_per_second=153.554Mi/s
FromMsgpack/small_signed_ints   61281736 ns     61260583 ns           12 bytes_per_second=76.2961Mi/s
```